Refactor + removing warnings

+++ src/safety_zone/int_edges_visualization.cpp
@@ -41,7 +41,7 @@ void IntEdgesVisualization::init() {
   prism_->subscribe(this);
 
   auto polygon = prism_->getPolygon().outer();
-  for(int i=0; i<polygon.size() - 1; i++){
+  for(size_t i=0; i<polygon.size() - 1; i++){
     addEdgeIntMarker(polygon[i], polygon[i+1], prism_->getMaxZ(), prism_->getMinZ(), i);
   }
   update();
@@ -88,13 +88,13 @@ void IntEdgesVisualization::update() {
 
   // Adding not present vertices
   if(polygon.size() - 1 > upper_names_.size()){
-    for(int i=upper_names_.size(); i<polygon.size() - 1; i++){
+    for(size_t i=upper_names_.size(); i<polygon.size() - 1; i++){
       addEdgeIntMarker(polygon[i], polygon[i+1], prism_->getMaxZ(), prism_->getMinZ(), i);
     }
   }
   
   // Updating the positions
-  for(int i=0; i<polygon.size() - 1; i++){
+  for(size_t i=0; i<polygon.size() - 1; i++){
     gm::Pose pose;
     pose.position.x = polygon[i].get<0>();
     pose.position.y = polygon[i].get<1>();
